move extension in extension folder
